feat(Programs): Use extensible transition array
Gone are the days of fixed upper limits on the number of transitions of
programs -- now only the sky (or your machine's memory, or size of your
address space, or your ulimit, or ... well, you get the idea) are the
limit.
Starts off with a small array for transitions (100), and doubles when
hitting maximal capacity.