fix(Termination): Fix logic for checking strict decrease of lex. rank functions
This has been wrong since the TACAS'13 paper, and only came up through the
certification effort, reported by Sebastiaan Joosten and Rene Thiemann.
Roughly, the bug is that our definition of strict & bounded decrease for
tuples was wrong. For example, for a lex. RF (f1, f2) with bounds (b1, b2),
we used the following definition:
 program well founded
  iff
 forall s, s' with s' post-state of s,
     (f1(s) > f1(s') && f1(s) >= b1)
  || (f1(s) >= f1(s') && f2(s) > f2(s') && f2(s) >= b2)

However, this is clearly satisfied for the following infinite reduction:
 (-1,1) >= (-2,2) > (-2,1) >= (-3,2) > (-3,1) >= (-4,2) > (-4,1) >= ...

It seems that this bug did not occur in practice (because the synthesis
procedure avoided generation RFs with this issue), but the check was
still wrong.